(PUP-7308) Propagate class tags to admissible whits
Without this patch applied, the admissible whit does not have the same
tags as the initial class. When sending a refresh event to a class and
running puppet with a specific tag, if that class loses it's tags,
then the agent unschedules all the refresh events (because it thinks
the class does not have the tag in question).

This patch copies tags from the initial class resource to the
admissible whit that it gets replaced with.